Split topology script
