+++ charmcraft.yaml
@@ -29,6 +29,13 @@ parts:
       craftctl default
     charm-strict-dependencies: true
     charm-requirements: [requirements.txt]
+  files:
+    plugin: dump
+    source: .
+    prime:
+      - LICENSE
+      - scripts
+      - templates
   libpq:
     build-packages:
       - libpq-dev
