+++ src/cluster_topology_observer.py
@@ -7,15 +7,10 @@ import logging
 import os
 import signal
 import subprocess
-import sys
-from time import sleep
 
-import requests
 from ops.charm import CharmBase, CharmEvents
 from ops.framework import EventBase, EventSource, Object
 from ops.model import ActiveStatus
-
-from constants import API_REQUEST_TIMEOUT, PATRONI_CLUSTER_STATUS_ENDPOINT
 
 logger = logging.getLogger(__name__)
 
@@ -79,7 +74,7 @@ class ClusterTopologyObserver(Object):
         pid = subprocess.Popen(  # noqa: S603
             [
                 "/usr/bin/python3",
-                "src/cluster_topology_observer.py",
+                "scripts/cluster_topology_observer.py",
                 self._charm._patroni._patroni_url,
                 f"{self._charm._patroni.verify}",
                 self._run_cmd,
@@ -118,45 +113,3 @@ class ClusterTopologyObserver(Object):
         """Juju-style tag identifying the unit being run by this charm."""
         unit_num = self._charm.unit.name.split("/")[-1]
         return f"unit-{self._charm.app.name}-{unit_num}"
-
-
-def dispatch(run_cmd, unit, charm_dir):
-    """Use the input juju-run command to dispatch a :class:`ClusterTopologyChangeEvent`."""
-    dispatch_sub_cmd = "JUJU_DISPATCH_PATH=hooks/cluster_topology_change {}/dispatch"
-    # Input is generated by the charm
-    subprocess.run([run_cmd, "-u", unit, dispatch_sub_cmd.format(charm_dir)])  # noqa: S603
-
-
-def main():
-    """Main watch and dispatch loop.
-
-    Watch the Patroni API cluster info. When changes are detected, dispatch the change event.
-    """
-    patroni_url, verify, run_cmd, unit, charm_dir = sys.argv[1:]
-
-    previous_cluster_topology = {}
-    while True:
-        cluster_status = requests.get(
-            f"{patroni_url}/{PATRONI_CLUSTER_STATUS_ENDPOINT}",
-            verify=verify,
-            timeout=API_REQUEST_TIMEOUT,
-        )
-        current_cluster_topology = {
-            member["name"]: member["role"] for member in cluster_status.json()["members"]
-        }
-
-        # If it's the first time the cluster topology was retrieved, then store it and use
-        # it for subsequent checks.
-        if not previous_cluster_topology:
-            previous_cluster_topology = current_cluster_topology
-        # If the cluster topology changed, dispatch a charm event to handle this change.
-        elif current_cluster_topology != previous_cluster_topology:
-            previous_cluster_topology = current_cluster_topology
-            dispatch(run_cmd, unit, charm_dir)
-
-        # Wait some time before checking again for a cluster topology change.
-        sleep(30)
-
-
-if __name__ == "__main__":
-    main()
