+++ tox.ini
@@ -7,8 +7,9 @@ env_list = lint, unit
 
 [vars]
 src_path = "{tox_root}/src"
+scripts_path = "{tox_root}/scripts"
 tests_path = "{tox_root}/tests"
-all_path = {[vars]src_path} {[vars]tests_path}
+all_path = {[vars]src_path} {[vars]scripts_path} {[vars]tests_path}
 
 [testenv]
 set_env =
@@ -47,7 +48,7 @@ set_env =
 commands_pre =
     poetry install --only main,charm-libs,unit --no-root
 commands =
-    poetry run coverage run --source={[vars]src_path} \
+    poetry run coverage run --source={[vars]src_path},{[vars]scripts_path} \
         -m pytest -v --tb native -s {posargs} {[vars]tests_path}/unit
     poetry run coverage report
     poetry run coverage xml
